image-service/requirements.txt
alembic==1.11.1
anyio==3.7.1
certifi==2023.5.7
click==8.1.4
colorama==0.4.6
fastapi==0.99.1
greenlet==2.0.2
h11==0.14.0
httptools==0.5.0
idna==3.4
Mako==1.2.4
MarkupSafe==2.1.3
minio==7.1.15
Pillow==10.0.0
psycopg2==2.9.6
pydantic==1.10.11
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0
sniffio==1.3.0
SQLAlchemy==2.0.18
starlette==0.27.0
typing_extensions==4.7.1
urllib3==2.0.3
uvicorn==0.22.0
watchfiles==0.19.0
websockets==11.0.3
